[Fix] Read memory output format on test of util/linux.
Previously uint8_t were displayed as chars and now they are displayed as
unsigned ints.


Former-commit-id: 2b9b7626566e3da608e79402ac7f379e826ac6e6

+++ unisim_lib/unisim/util/os/linux/test.cc
@@ -126,7 +126,7 @@ bool MemInterface::ReadMemory(uint32_t addr, uint8_t * const buffer,
   }
   std::cout << "            --->" << std::hex;
   for ( uint32_t index = 0; index < size; ++index)
-    std::cout << " " << buffer[index];
+    std::cout << " " << (unsigned int)buffer[index];
   std::cout << std::dec << std::endl;
   return true;
 }
